Implement integration tests for signup use case.

+++ infra/src/routes/accounts.rs
@@ -1,16 +1,17 @@
 use actix_web::http::StatusCode;
 use actix_web::{web, HttpResponse};
-use domain::models::passwords::RawPassword;
 use secrecy::SecretString;
 use time::OffsetDateTime;
 use uuid::Uuid;
 
+use domain::models::passwords::RawPassword;
 use domain::models::primitives::*;
 use domain::models::user::User;
+use macros::{Builder, Getter};
 use use_cases::accounts::{SignupUser, SignupUserBuilder};
 use use_cases::UseCaseError;
 
-use crate::routes::common::{ErrorResponseBody, ProcessRequestError, ProcessRequestResult};
+use crate::routes::{ErrorResponseBody, ProcessRequestError, ProcessRequestResult};
 use crate::RequestContext;
 
 /// アカウントスコープを返却する。
@@ -36,42 +37,55 @@ pub async fn signup(
 /// ```json
 /// {"email": "foo@example.com", "password": "p@ssw0rd", "familyName": "Yamada", "givenName": "Taro", "postalCode": "899-7103", "address": "鹿児島県志布志市志布志町志布志2-1-1", "fixedPhoneNumber": "099-472-1111", "mobilePhoneNumber": "090-1234-5678", "remarks": "日本に実際に存在するややこしい地名です。"}
 /// ```
-#[derive(Debug, Clone, serde::Deserialize)]
+#[derive(Debug, Clone, serde::Deserialize, Getter, Builder)]
 #[serde(rename_all = "camelCase")]
 pub struct SignupRequestBody {
     /// Eメールアドレス
+    #[getter(ret = "ref")]
     email: String,
     /// 未加工なパスワード
+    #[getter(ret = "ref")]
     password: SecretString,
     /// 苗字
+    #[getter(ret = "ref")]
     family_name: String,
     /// 名前
+    #[getter(ret = "ref")]
     given_name: String,
     /// 郵便番号
+    #[getter(ret = "ref")]
     postal_code: String,
     /// 住所
+    #[getter(ret = "ref")]
     address: String,
     /// 固定電話番号
+    #[getter(ret = "ref")]
     fixed_phone_number: Option<String>,
     /// 携帯電話番号
+    #[getter(ret = "ref")]
     mobile_phone_number: Option<String>,
     /// 備考
+    #[getter(ret = "ref")]
     remarks: Option<String>,
 }
 
 /// サインアップ・レスポンス・ボディ
-#[derive(Debug, Clone, serde::Serialize)]
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, Getter)]
 #[serde(rename_all = "camelCase")]
 pub struct SignupResponseBody {
     /// ユーザーID
+    #[getter(ret = "val")]
     id: Uuid,
     /// Eメール・アドレス
+    #[getter(ret = "ref")]
     email: String,
     /// 作成日時
     #[serde(with = "time::serde::rfc3339")]
+    #[getter(ret = "val")]
     created_at: OffsetDateTime,
     /// 更新日時
     #[serde(with = "time::serde::rfc3339")]
+    #[getter(ret = "val")]
     updated_at: OffsetDateTime,
 }
 
@@ -120,7 +134,7 @@ impl From<User> for SignupResponseBody {
 impl From<UseCaseError> for ProcessRequestError {
     fn from(value: UseCaseError) -> Self {
         let body = ErrorResponseBody {
-            error_code: Some(value.error_code),
+            error_code: Some(value.error_code as u32),
             message: value.message,
         };
         match value.kind {
